# Conflicts:
#	ambari-web/app/styles/application.less
#	ambari-web/app/styles/wizard.less
#	ambari-web/app/templates/installer.hbs

+++ ambari-server/src/main/java/org/apache/ambari/server/orm/DBAccessorImpl.java
@@ -956,8 +956,12 @@ public class DBAccessorImpl implements DBAccessor {
 
   @Override
   public void dropTable(String tableName) throws SQLException {
-    String query = dbmsHelper.getDropTableStatement(tableName);
-    executeQuery(query);
+    if (tableExists(tableName)){
+      String query = dbmsHelper.getDropTableStatement(tableName);
+      executeQuery(query);
+    } else {
+      LOG.warn("{} table doesn't exists, skipping", tableName);
+    }
   }
 
 
@@ -1551,8 +1555,12 @@ public class DBAccessorImpl implements DBAccessor {
    */
   @Override
   public void clearTable(String tableName) throws SQLException {
-    String sqlQuery = "DELETE FROM " + convertObjectName(tableName);
-    executeQuery(sqlQuery);
+    if (tableExists(tableName)){
+      String sqlQuery = "DELETE FROM " + convertObjectName(tableName);
+      executeQuery(sqlQuery);
+    } else {
+      LOG.warn("{} table doesn't exists, skipping", tableName);
+    }
   }
 
   /**
@@ -1564,7 +1572,11 @@ public class DBAccessorImpl implements DBAccessor {
    */
   @Override
   public void clearTableColumn(String tableName, String columnName, Object value) throws SQLException {
-    String sqlQuery = String.format("UPDATE %s SET %s = ?", convertObjectName(tableName), convertObjectName(columnName));
-    executePreparedUpdate(sqlQuery, value);
+    if (tableExists(tableName)){
+      String sqlQuery = String.format("UPDATE %s SET %s = ?", convertObjectName(tableName), convertObjectName(columnName));
+      executePreparedUpdate(sqlQuery, value);
+    } else {
+      LOG.warn("{} table doesn't exists, skipping", tableName);
+    }
   }
 }
